Refactor: move database.py and models.py to src/
